+++ DatadogCore/Tests/Datadog/Logs/CrashLogReceiverTests.swift
@@ -145,7 +145,14 @@ class CrashLogReceiverTests: XCTestCase {
             applicationBuildNumber: crashContext.buildNumber,
             buildId: nil,
             variant: core.context.variant,
-            dd: .init(device: .init(architecture: crashContext.device.architecture)),
+            dd: .init(
+                device: .init(
+                    brand: crashContext.device.brand,
+                    name: crashContext.device.name,
+                    model: crashContext.device.model,
+                    architecture: crashContext.device.architecture
+                )
+            ),
             os: .init(
                 name: crashContext.device.osName,
                 version: crashContext.device.osVersion,